TST/REF: test_sample
